pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "automatic-qa-categories"
version = "0.1.0"
description = "Local hybrid QA audit pipeline for customer-support agent replies"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "pydantic>=2.7,<3",
  "PyYAML>=6.0.1,<7",
  "requests>=2.31,<3",
  "python-dotenv>=1.0,<2"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0,<9",
  "jsonschema>=4.22,<5"
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
addopts = "-q"

[tool.ruff]
line-length = 100
target-version = "py311"
